+++ src/oms_sensemaking/service.py
@@ -103,6 +103,7 @@ def create_app(config: Settings) -> FastAPI:
         version=__version__,
         lifespan=lifespan,
         root_path=config.root_path,
+        docs_url="/",
     )
 
     # initialize gzip middleware