+++ tests/inference/rules/test_incursions.py
@@ -4,7 +4,6 @@ from unittest.mock import MagicMock
 import pytest
 from oms_sdk import DEFAULT_ACM
 from oms_sdk.generated.generated_graphql_client import (
-    ActivityState,
     AttributeAttribute,
     AttributeQuery,
     AttributeType,
@@ -261,7 +260,7 @@ def test_new_incursion_region1(
     rule.action(RuleContext(observation=observational_node_region1))
     mock_get_attributes.assert_called_with(
         AttributeQuery(
-            attributeIri=SETTINGS.inference_incursion_attribute_iri,
+            attributeIris=[SETTINGS.inference_incursion_attribute_iri],
             attributeValue=StringQuery(equals="Incursion"),
             attributeType={"is": AttributeType.GEOSPATIAL},
             geometry=GeoQuery(queryGeoJson=areas_of_interest[0]),
@@ -291,7 +290,7 @@ def test_new_incursion_region1(
             classIri=SETTINGS.inference_incursion_class_iri,
             name="Incursion",
             description=f"Incursion detected into {areas_of_interest[0]}",
-            state=ActivityState.UNKNOWN,
+            state=SETTINGS.inference_incursion_activity_state,
             nodeId=observational_node_region1.nodeId,
             observationIds=[observational_node_region1.id],
             startTime=observational_node_region1.startTime,
@@ -315,7 +314,7 @@ def test_new_incursion_region2(
     rule.action(RuleContext(observation=observational_node_region2))
     mock_get_attributes.assert_called_with(
         AttributeQuery(
-            attributeIri=SETTINGS.inference_incursion_attribute_iri,
+            attributeIris=[SETTINGS.inference_incursion_attribute_iri],
             attributeValue=StringQuery(equals="Incursion"),
             attributeType={"is": AttributeType.GEOSPATIAL},
             geometry=GeoQuery(queryGeoJson=areas_of_interest[1]),
@@ -345,7 +344,7 @@ def test_new_incursion_region2(
             classIri=SETTINGS.inference_incursion_class_iri,
             name="Incursion",
             description=f"Incursion detected into {areas_of_interest[1]}",
-            state=ActivityState.UNKNOWN,
+            state=SETTINGS.inference_incursion_activity_state,
             nodeId=observational_node_region2.nodeId,
             observationIds=[observational_node_region2.id],
             startTime=observational_node_region2.startTime,
@@ -378,7 +377,7 @@ def test_two_existing_incursions(
     rule.action(RuleContext(observation=observational_node_region1))
     mock_get_attributes.assert_called_with(
         AttributeQuery(
-            attributeIri=SETTINGS.inference_incursion_attribute_iri,
+            attributeIris=[SETTINGS.inference_incursion_attribute_iri],
             attributeValue=StringQuery(equals="Incursion"),
             attributeType={"is": AttributeType.GEOSPATIAL},
             geometry=GeoQuery(queryGeoJson=areas_of_interest[0]),